% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_functions.R
\docType{package}
\name{spell.replacer-package}
\alias{spell.replacer}
\alias{spell.replacer-package}
\title{spell.replacer: Probabilistic Spelling Correction}
\description{
This package provides functions for automatic spelling correction in
character vectors using probabilistic methods based on string distance and
word frequency data from the Corpus of Contemporary American English (COCA).
}
\author{
\strong{Maintainer}: David Brown \email{dwb2@andrew.cmu.edu} (\href{https://orcid.org/0000-0001-7745-6354}{ORCID})

}
\keyword{internal}
